+++ src/components/editor/Space.svelte
@@ -9,19 +9,21 @@
 
     export let token: Token;
     export let space: string;
+    export let line: number | undefined;
     export let insertion: InsertionPoint | undefined = undefined;
+    export let first = false;
 
     $: insertionIndex =
         insertion !== undefined
             ? space.split('\n', insertion.line).join('\n').length + 1
             : undefined;
     // If there's an insertion, figure out what space to render before the insertion.
-    $: beforeSpaces =
+    $: beforeSpacesByLine =
         insertionIndex === undefined
             ? []
             : render(space.substring(0, insertionIndex), $spaceIndicator);
     // If there's no insertion, just render the space, otherwise render the right side of the insertion.
-    $: afterSpaces = render(
+    $: afterSpacesByLine = render(
         insertionIndex === undefined ? space : space.substring(insertionIndex),
         $spaceIndicator,
     );
@@ -33,6 +35,12 @@
                 : text.replaceAll(' ', '\xa0').replaceAll('\t', TAB_TEXT)
         ).split('\n');
     }
+
+    $: firstLine =
+        line !== undefined
+            ? line - beforeSpacesByLine.length - afterSpacesByLine.length + 1
+            : undefined;
+    $: showLines = true;
 </script>
 
 <!-- 
@@ -42,23 +50,43 @@
 -->
 {#key $spaceIndicator}
     {#key space}
-        <span class="space" role="none" data-id={token.id} data-uiid="space"
-            ><span role="none" class="before"
-                >{#each beforeSpaces as s, index}{#if index > 0}<span
-                            ><br class="break" /></span
-                        >{/if}{#if s === ''}&ZeroWidthSpace;{:else}<span
-                            class="line"
-                            data-uiid="space-text">{s}</span
-                        >{/if}{:else}&ZeroWidthSpace;{/each}{#if insertion}<InsertionPointView
-                    />{/if}</span
-            ><span role="none" class="after"
-                >{#each afterSpaces as s, index}{#if index > 0}<span
-                            ><br class="break" /></span
-                        >{/if}<span class="line" data-uiid="space-text"
-                        >{s}</span
-                    >{/each}</span
-            ></span
-        >
+        {#key line}
+            <span
+                class="space"
+                role="none"
+                data-id={token.id}
+                data-uiid="space"
+            >
+                <span role="none" class="before"
+                    >{#if first && showLines}<div class="line-number">1</div
+                        >{/if}{#each beforeSpacesByLine as s, index}{#if index > 0}<span
+                                ><br
+                                    class="break"
+                                />{#if showLines && firstLine !== undefined}<div
+                                        class="line-number"
+                                        >{firstLine + index}</div
+                                    >{/if}</span
+                            >{/if}{#if s === ''}&ZeroWidthSpace;{:else}<span
+                                class="line"
+                                data-uiid="space-text">{s}</span
+                            >{/if}{:else}&ZeroWidthSpace;{/each}{#if insertion}<InsertionPointView
+                        />{/if}</span
+                ><span role="none" class="after"
+                    >{#each afterSpacesByLine as s, index}{#if index > 0}<span
+                                ><br
+                                    class="break"
+                                />{#if showLines && firstLine !== undefined}<div
+                                        class="line-number"
+                                        >{firstLine +
+                                            beforeSpacesByLine.length +
+                                            index}</div
+                                    >{/if}</span
+                            >{/if}<span class="line" data-uiid="space-text"
+                            >{s}</span
+                        >{/each}</span
+                ></span
+            >
+        {/key}
     {/key}
 {/key}
 
@@ -72,5 +100,13 @@
     /* If the space is in something dragged, hide it */
     :global(.dragged) .space {
         visibility: hidden;
+    }
+
+    .line-number {
+        display: inline-block;
+        width: calc((var(--line-count)) * 1em);
+        font-size: var(--wordplay-small-font-size);
+        vertical-align: middle;
+        color: var(--wordplay-inactive-color);
     }
 </style>